+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx118-en-CA.html
@@ -41,7 +41,7 @@
         <p>Click the {% include 'firefox/whatsnew/includes/fx118/translate-icon.svg' %} icon in Firefox’s address bar and enjoy content across the web.</p>
 
         <p>
-          <a id="cta-now" class="mzp-c-button mzp-t-product" href="https://www.mozilla.org/es-MX/" data-cta-type="button" data-cta-text="Try it now">Try it now</a>
+          <a id="cta-now" class="mzp-c-button mzp-t-product" href="https://www.mozilla.org/fr/" data-cta-type="button" data-cta-text="Try it now">Try it now</a>
         </p>
 
     </div>
