+++ tests/pages/firefox/whatsnew/whatsnew_118_na.py
@@ -1,0 +1,18 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+from selenium.webdriver.common.by import By
+
+from pages.base import BasePage
+
+
+class FirefoxWhatsNew118NAPage(BasePage):
+    _URL_TEMPLATE = "/{locale}/firefox/118.0/whatsnew/"
+
+    # NA pages
+    _set_try_it_button_locator = (By.ID, "cta-now")
+
+    @property
+    def is_try_it_button_displayed(self):
+        return self.is_element_displayed(*self._set_try_it_button_locator)
